bloomfilter: move internal functions into Classic.Internal
The same as what we did for MBloom, but now for the immutable Bloom
type and functions.

So now the Bloom representation and functions that need access to it
live in Classic.Internal while functions that don't need representation
access live in the .Classic top level module.
