+++ bloomfilter/src/Data/BloomFilter/Classic.hs
@@ -70,19 +70,15 @@ module Data.BloomFilter.Classic (
     unsafeFreeze,
 ) where
 
-import           Control.Exception (assert)
 import           Control.Monad.Primitive (PrimMonad, PrimState, RealWorld,
                      stToPrim)
 import           Control.Monad.ST (ST, runST)
-import           Data.Primitive.ByteArray (ByteArray, MutableByteArray)
-import           Data.Word (Word64)
+import           Data.Primitive.ByteArray (MutableByteArray)
 
-import qualified Data.BloomFilter.Classic.BitVec64 as V
 import           Data.BloomFilter.Classic.Calc
 import           Data.BloomFilter.Classic.Internal hiding (deserialise)
 import qualified Data.BloomFilter.Classic.Internal as Internal
-import           Data.BloomFilter.Hash (CheapHashes, Hash, Hashable, evalHashes,
-                     makeHashes)
+import           Data.BloomFilter.Hash (CheapHashes, Hash, Hashable, makeHashes)
 
 import           Prelude hiding (elem, notElem, read)
 
@@ -113,66 +109,11 @@ create bloomsize body =
 insert :: Hashable a => MBloom s a -> a -> ST s ()
 insert !mb !x = insertHashes mb (makeHashes x)
 
--- | Create an immutable Bloom filter from a mutable one.  The mutable
--- filter may be modified afterwards.
-freeze :: MBloom s a -> ST s (Bloom a)
-freeze MBloom { mbNumBits, mbNumHashes, mbBitArray } = do
-    bitArray <- V.freeze mbBitArray
-    let !bf = Bloom {
-                numBits   = mbNumBits,
-                numHashes = mbNumHashes,
-                bitArray
-              }
-    assert (bloomInvariant bf) $ pure bf
-
--- | Create an immutable Bloom filter from a mutable one.  The mutable
--- filter /must not/ be modified afterwards, or a runtime crash may
--- occur.  For a safer creation interface, use 'freeze' or 'create'.
-unsafeFreeze :: MBloom s a -> ST s (Bloom a)
-unsafeFreeze MBloom { mbNumBits, mbNumHashes, mbBitArray } = do
-    bitArray <- V.unsafeFreeze mbBitArray
-    let !bf = Bloom {
-                numBits   = mbNumBits,
-                numHashes = mbNumHashes,
-                bitArray
-              }
-    assert (bloomInvariant bf) $ pure bf
-
--- | Copy an immutable Bloom filter to create a mutable one.  There is
--- no non-copying equivalent.
-thaw :: Bloom a -> ST s (MBloom s a)
-thaw Bloom { numBits, numHashes, bitArray } = do
-    mbBitArray <- V.thaw bitArray
-    pure MBloom {
-      mbNumBits   = numBits,
-      mbNumHashes = numHashes,
-      mbBitArray
-    }
-
 -- | Query an immutable Bloom filter for membership.  If the value is
 -- present, return @True@.  If the value is not present, there is
 -- /still/ some possibility that @True@ will be returned.
 elem :: Hashable a => a -> Bloom a -> Bool
 elem elt ub = elemHashes (makeHashes elt) ub
-
--- | Query an immutable Bloom filter for membership using already constructed 'Hashes' value.
-elemHashes :: CheapHashes a -> Bloom a -> Bool
-elemHashes !ch Bloom { numBits, numHashes, bitArray } =
-    go 0
-  where
-    go :: Int -> Bool
-    go !i | i >= numHashes
-          = True
-    go !i = let idx' :: Word64
-                !idx' = evalHashes ch i in
-            let idx :: Int
-                !idx = fromIntegral (idx' `V.unsafeRemWord64` fromIntegral numBits) in
-            -- While the idx' can cover the full Word64 range,
-            -- after taking the remainder, it now must fit in
-            -- and Int because it's less than the filter size.
-            if V.unsafeIndex bitArray idx
-              then go (i + 1)
-              else False
 
 -- | Query an immutable Bloom filter for non-membership.  If the value
 -- /is/ present, return @False@.  If the value is not present, there
@@ -189,14 +130,6 @@ notElemHashes !ch !ub = not (elemHashes ch ub)
 -- /still/ some possibility that @True@ will be returned.
 read :: Hashable a => a -> MBloom s a -> ST s Bool
 read elt mb = readHashes (makeHashes elt) mb
-
--- | Return the size of the Bloom filter.
-size :: Bloom a -> BloomSize
-size Bloom { numBits, numHashes } =
-    BloomSize {
-      sizeBits   = numBits,
-      sizeHashes = numHashes
-    }
 
 -- | Build an immutable Bloom filter from a seed value.  The seeding
 -- function populates the filter as follows.
@@ -236,12 +169,6 @@ fromList policy xs =
     create bsize (\b -> mapM_ (insert b) xs)
   where
     bsize = sizeForPolicy policy (length xs)
-
-serialise :: Bloom a -> (BloomSize, ByteArray, Int, Int)
-serialise b@Bloom{bitArray} =
-    (size b, ba, off, len)
-  where
-    (ba, off, len) = V.serialise bitArray
 
 {-# SPECIALISE deserialise :: BloomSize
                            -> (MutableByteArray RealWorld -> Int -> Int -> IO ())
